pyproject.toml
[project]
name = "semanticsearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "selectolax>=0.3.28",
    "torch>=2.6.0",
    "transformers>=4.49.0",
    "tqdm>=4.67.1",
]
